There were several wildcard imports which were part of the dependency problem.
All classes were already put into camplight.api so I decided to break
compatibility and use camplight.api instead of camplight.

Users using the command line tool will not be affected by this change.
It only affects users of the Python API. But the change is small so I
think it's appropriate.

Bumped to version 0.5.